testvectors: restructure aead vectors into modules
Splits up the aead vectors into their own module(s). We want to
add more vectors for different AEADs, and having them all in lib.rs
would get unwieldy.
